requirements.txt
begins
fs
pyserial
six
begins